pyproject.toml
[tool.poetry]
name = "sber-fastapi"
version = "0.1.0"
description = ""
authors = ["Ivylou <princequincyy@gmail.com>"]
readme = "README.md"
packages = [{include = "sber_fastapi"}]

[tool.poetry.dependencies]
python = "3.8.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.12"
psycopg2-binary = "^2.9.6"
alembic = "^1.10.4"
pydantic = {extras = ["email"], version = "^1.10.7"}
tenacity = "^8.2.2"
python-dotenv = "^1.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cffi = "^1.15.1"
python-multipart = "^0.0.6"
pytest = "^7.3.1"
httpx = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
